+++ frontend/src/hooks/tanstack-query/station-details/reports/useUpdateStationChargerReport.ts
@@ -2,8 +2,8 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { getLocalStorage } from '@utils/storage';
 
+import { serverStore } from '@stores/config/serverStore';
 import { modalActions } from '@stores/modalStore';
-import { serverStore } from '@stores/serverStore';
 
 import { DEFAULT_TOKEN, SERVERS } from '@constants';
 import { LOCAL_KEY_TOKEN } from '@constants/storageKeys';
