+++ frontend/src/hooks/tanstack-query/useSearchedStations.ts
@@ -2,8 +2,8 @@ import { useQuery } from '@tanstack/react-query';
 
 import { useExternalValue } from '@utils/external-state';
 
+import { serverStore } from '@stores/config/serverStore';
 import { searchWordStore } from '@stores/searchWordStore';
-import { serverStore } from '@stores/serverStore';
 
 import { SERVERS } from '@constants';
 import { ERROR_MESSAGES } from '@constants/errorMessages';
